requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
huggingface_hub
gradio
requests
sentence-transformers
faiss-cpu
markdown